# backend/app/docs/openapi.yaml
openapi: "3.1.0"

# ─────────────────────────────────────────
# Info 
# ─────────────────────────────────────────
info:
  title: Disney Movies API - OpenAPI
  version: "1.0.0"
  description: | 
     API REST para consultar y analizar el dataset de películas Disney (1937–2016)

      ## Endpoints
    
      ## Endpoints
      | Método | Ruta                  | Descripción                        |
      |--------|-----------------------|------------------------------------|
      | GET    | `/api/v1/movies`      | Lista paginada de películas        |
      | GET    | `/api/v1/movies/search` | Filtrar por género y/o rating    |
      | GET    | `/api/v1/movies/stats`  | Estadísticas del dataset         |
      | GET    | `/api/v1/movies/{title}` | Buscar película por título      |

      ## Ratings disponibles
      `G` · `PG` · `PG-13` · `R` · `Not Rated`

  contact:
    name: Disney Movies API's
    url: http://localhost:5000/docs

  license:
    name: MIT


# ─────────────────────────────────────────
# SERVERS
# ─────────────────────────────────────────
servers:
  - url: http://localhost:5000
    description: Desarrollo local con Docker & Nginx Proxy
  - url: http://staging.api.example.com
    description: Staging

# ─────────────────────────────────────────
# TAGS
# ─────────────────────────────────────────
tags:
  - name: Health
    description: Estado del Servidor
  - name: Movies
    description: Operaciones | Endpoints sobre el dataset de películas de Disney.

# ─────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────
paths:

  # - Init Path | Health - #
  /:
    get:
      tags: [Health]
      summary: Health check
      operationId: healthCheck
      responses: 
        "200":
          description: Servidor Activo
          content: 
            application/json:
              schema: 
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
                message: Disney Movies API's Running
                docs: /docs
                openapi: /openapi.json



  # =================== List Movies [Get Movies] =======================
  /api/v1/movies:
    get:
      tags: [Movies]
      summary: List Of Movies
      operationId: listMovies
      description: |
        Return all list of movies Disney. Dataset Disney movies
        Use "limit" & "offset" to paginated
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200": 
          description: "List Of Movies"
          content:
            application/jason:
              schema:
                $ref: "#/components/schema/MovieListResponse"
              example: 
                count: 2 
                total: 579 
                limit: 2
                offset: 0 
                data:
                  - movie_title: Snow White and the Seven Dwarfs 
                    release_date: "21/12/1937" 
                    genre: Musical
                    rating: G
                    total_gross: 184925485
                    adjusted_gross: 5228953251
                  - movie_title: Pinocchio
                    release_date: "09/02/1940" 
                    genre: Adventure
                    rating: G 
                    total_gross: 84300000,
                    adjusted_gross: 2188229052
        "500":
          $ref: "#/components/InternalError"

  # =================== Movies for Title [Search] =======================
# ── Stats ─────────────────────────────
  /api/v1/movies/stats:
    get:
      tags: [Movies]
      summary: Dataset Statics (stats)
      operationId: getStats
      description: |
        Return Statics of the Movies
      responses:
        "200":
          description: Dataset General Statics (stats)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"
              example:
                total_movies: 579
                genres:
                  - genre: Action
                    count: 32
                    total_gross_sum: 2341000000
                    avg_gross: 73156250.0
                  - genre: Adventure
                    count: 141
                    total_gross_sum: 12870000000
                    avg_gross: 91276595.0
                top_grossing:
                  movie_title: Star Wars Ep. VII The Force Awakens
                  release_date: "18/12/2015"
                  genre: Adventure
                  rating: PG-13
                  total_gross: 936662225
                  adjusted_gross: 936662225
                most_recent:
                  movie_title: "Rogue One: A Star Wars Story"
                  release_date: "16/12/2016"
                  genre: Adventure
                  rating: PG-13
                  total_gross: 529483936
                  adjusted_gross: 529483936
        "500":
          $ref: "#/components/responses/InternalError"
 
 
  # ── Get By Title ──────────────────────
  /api/v1/movies/{title}:
    get:
      tags: [Movies]
      summary: Get Movies for Title
      operationId: getMovieByTitle
      description: |
        Search Movies for Title & Get Movies
      parameters:
        - name: title
          in: path
          required: true
          description: Title for movie
          schema:
            type: string
          example: The Lion King
      responses:
        "200":
          description: Película encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieResponse"
              example:
                movie_title: The Lion King
                release_date: "15/06/1994"
                genre: Adventure
                rating: G
                total_gross: 422780140
                adjusted_gross: 761640898
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"


# ─────────────────────────────────────────
# === COMPONENTS (Componentes) ===
# ─────────────────────────────────────────
components:

## === Parameters === ## 
  parameters:

    LimitParam:
      name: limit
      in: query
      required: false
      description: Number the Movies per Page (Paginated - max. 100)
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100

    OffsetParam: 
      name: offset
      in: query
      required: false
      description: Number to Register to skip(Omitir)"
      schema:
        type: integer
        default: 0
        minimum: 0


## === Paginated List Movies === ##
  schemas:
    
    # === Individual Movie === #
    MovieResponse:
      type: object
      required: 
        - movie_title
        - release_date
        - genre
        - rating 
        - total_gross
        - adjusted_gross
      properties: 
        movie_title:
          type: string
          description: Title for movie
          example: The Lion King
        release_date: 
          type: string
          description: Date of the movie premier (DD/MM/YYYY)
          example: "15/06/1994"
        genre: 
          type: string
          description: Genre of the movie
          example: Adventure
        rating: 
          type: string
          description: MPAA Clasification
          example: [G, PG, PG-13, R, "Not Rated"]
        total_gross: 
          type: integer
          description: Total gross movie in USD
          minimum: 0
          example: 422780140
        adjusted_gross:
          type: integer
          description: Adjusted gross movie for Inflation in USD
          minimum: 0
          example: 761640898

    # === Movie Paginated === #
    MovieListResponse:
      type: object
      required: [count, total, limit, offset, data]
      properties: 
        count: 
          type: integer
          description: Number of movies on this response
          example: 10
        total: 
          type: integer
          description: Movie Total in the Dataset (DB)
          example: 579
        limit: 
          type: integer
          description: Limit in the request of the movies
          example: 10
        offset: 
          type: integer
          description: Offset of Pagination
          example: 0
        data: 
          type: array
          items: 
            $ref: "#/components/schemas/MovieResponse"
            

    # - Health -
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
        message: 
          type: string
          example: Disney Movies APIs running
        docs: 
          type: string
          example: /docs
        openapi:
          type: string
          example: /openapi.json


    # — Error —
    ErrorResponse:
      type: object
      required: [detail, code]
      properties:
        detail:
          type: string
          example: Movie 'Bambi' not found.
        code:
          type: integer
          example: 404


# === Response === #
  responses: 
  
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            detail: "Movie 'Bambi' not found."
            code: 404

    InternalError:
      description: Descripción del error
      content: 
        application/jason:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example: 
            details: "csv not found at path /app/data/disney_movies_2_cleaned.csv"
            code: 500